requirements.txt
sqlite3
python-dotenv
discord.py